+++ src/preset/bidirectional.rs
@@ -30,3 +30,41 @@ impl<P: Bundle, N: Bundle> SpawnableList<BindingOf> for Bidirectional<P, N> {
         2
     }
 }
+
+impl Bidirectional<Binding, Binding> {
+    /// Maps A and D keys as 1-dimensional input.
+    #[must_use]
+    pub fn ad_keys() -> Self {
+        Self {
+            positive: KeyCode::KeyD.into(),
+            negative: KeyCode::KeyA.into(),
+        }
+    }
+
+    /// Maps W and S keys as 1-dimensional input.
+    #[must_use]
+    pub fn ws_keys() -> Self {
+        Self {
+            positive: KeyCode::KeyW.into(),
+            negative: KeyCode::KeyS.into(),
+        }
+    }
+
+    /// Maps left and right keyboard arrow keys as 1-dimensional input.
+    #[must_use]
+    pub fn horizontal_arrow_keys() -> Self {
+        Self {
+            positive: KeyCode::ArrowRight.into(),
+            negative: KeyCode::ArrowLeft.into(),
+        }
+    }
+
+    /// Maps up and down keyboard arrow keys as 1-dimensional input.
+    #[must_use]
+    pub fn vertical_arrow_keys() -> Self {
+        Self {
+            positive: KeyCode::ArrowRight.into(),
+            negative: KeyCode::ArrowLeft.into(),
+        }
+    }
+}